RUMM-2356 Add Snapshot Producer first layer of logic

+++ library/dd-sdk-android-session-replay/src/main/kotlin/com/datadog/android/sessionreplay/recorder/LongExt.kt
@@ -7,15 +7,15 @@
 package com.datadog.android.sessionreplay.recorder
 
 /**
- * Normalizes an Int value (font size, view dimension, view position, etc.) according with the
+ * Normalizes a Long value (font size, view dimension, view position, etc.) according with the
  * device pixels density.
  * Example: if a device has a DPI = 2, the normalized height of a view will be
  * view.height/2.
  * @param density
  */
-internal fun Int.densityNormalized(density: Float): Int {
+internal fun Long.densityNormalized(density: Float): Long {
     if (density == 0f) {
         return this
     }
-    return (this / density).toInt()
+    return (this / density).toLong()
 }
